refined home logo logic
